Consolidate javascript feature tests
